abstract out statusFunction
